deploy/server/requirements.txt
flask>=2.0.0
cryptography>=41.0.0
requests>=2.31.0
dnspython>=2.4.0
pycryptodome>=3.19.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
redis>=4.5.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0
PyJWT>=2.8.0
